added IgnoreInequality flag directly on InstantCast finder and fixed an ICD bug

+++ GW2EIEvtcParser/EIData/InstantCastFinders/BuffGainCastFinder.cs
@@ -26,7 +26,7 @@ namespace GW2EIEvtcParser.EIData
             var applies = combatData.GetBuffData(BuffID).OfType<BuffApplyEvent>().GroupBy(x => x.To).ToDictionary(x => x.Key, x => x.ToList());
             foreach (KeyValuePair<AgentItem, List<BuffApplyEvent>> pair in applies)
             {
-                long lastTime = long.MinValue;
+                long lastTime = int.MinValue;
                 foreach (BuffApplyEvent bae in pair.Value)
                 {
                     if (bae.Time - lastTime < ICD)
